test/nvmf: try to make reactors more busy in interrupt test
This test verifies that reactors are idle when there's no IO and that
they're idle when handling IO.  The latter check can be a problem,
because the business of a core when handling IOs varies and can never
reach the threshold that we use to consider a core to be busy.

To make that second check more reliable, let's do a couple of things:
increase queue depth to 256 and decrease busy threshold to 30.  Also,
while here, use double parenthesis for arithmetic comparisons.

Fixes: #3520

Signed-off-by: Konrad Sztyber <konrad.sztyber@intel.com>
Change-Id: I2937b46471d523bb496e7b3878f43c541047adaf
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24827
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>