test/common: Make sure get_nvmes_with_ns_management() doesn't report failure
This function is meant to simply list nvme devices which support
ns management, it shouldn't trigger errexit (it's up to the caller to
verify returned list of devices).

Currently, this function may report failure in case last ctrl from
the list doesn't have proper support enabled - this may be problematic
in case there's a mix of different nvmes installed with different
capabilities.

Change-Id: I57d96a1404534003c8d1a46bf85eda6e316da529
Signed-off-by: Michal Berger <michal.berger@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24834
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>