trace: do not define fn in SPDK_TRACE_REGISTER_FN
Newer lcov try to detect issues with function definitions, and throw
warnings if function definitions have overlapping lines. Since
SPDK_TRACE_REGISTER_FN defines both the constructor function and
the trace function, it treats this as an overlap.

Note this is the only one of our REGISTER functions in SPDK that
effectively defines two functions in the same macro.

So change the macro so that the user actually defines the
trace function instead.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: Ib9d0d1af7674c5ff9b65c12c1e4a5b05e6e110f5
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24836
Reviewed-by: Michal Berger <michal.berger@intel.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot