bdevperf: statistics calculation in separate function
The statistics calculation is moved out of performance_dump_job()
function to separate function (bdevperf_job_get_stats()). Now we can
reuse the statistics in two places: in a function that prints test
results on the console and in a function that provide test results in
RPC response.
ema_period was removed from bdevperf_stats and now is passed as a
variable.

Change-Id: I7d9b59a7d9cf24fd35cdd059a4ae9fe715028423
Signed-off-by: Marcin Spiewak <marcin.spiewak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24892
Community-CI: Mellanox Build Bot
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>