sock/posix: fix busy poller in the interrupt mode
Setting EPOLLOUT flag makes the epoll_wait exit immidiately on the
tqpair socket as it's always ready for the write operation.
This was a bad fix for the async writes in the interrupt mode. Proper
fix is to flush async writes on nvmf/tcp level.

Extend test script for the interrupt mode to include case of idle
initiator in which CPU cores shall not be busy looping.

Change-Id: I17a97a74be9740dd67531f6e3b72bf38f0850edd
Signed-off-by: Krzysztof Goreczny <krzysztof.goreczny@dell.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24959
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Reviewed-by: Jim Harris <jim.harris@samsung.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>