module/raid: Assign bdev_io buffers to raid_io
If an application does not allocate data and metadata
buffers for the read operation and relies on the generic
bdev layer to do the allocations, the data iovecs
and metadata pointer was not copied to the raid_io
and not pushed down the I/O stack.

This fix copies the pointers once the allocations are complete.
If the application did provide the buffers, this flow will
still work since those pointers are already in the bdev_io
and raid_io.

Change-Id: Ib9e212ec97b59c0ed57f2e48a35fdf8fd2a21a52
Signed-off-by: Slawomir Ptak <slawomir.ptak@intel.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/24961
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>