event: don't print core unlock warnings if cores were never locked
If we fail app init path, we try to unlock any cores that happened
to be locked. But the code path was only checking for -1, and by
default the core lock array would have 0 values. So only try to
unlock if the core lock fd is not -1 nor 0.

Signed-off-by: Jim Harris <jim.harris@samsung.com>
Change-Id: I3a2f7bb8bf11c5bc3d8c59b213a356aff87b30d0
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/25059
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>