bdev/nvme: Fix spdk_bdev_nvme_create()
This patch fixes spdk_bdev_nvme_create function by ensuring that
base_name received is persisted until we create the bdevs. Current all
in-tree callers persist the string parameter that is passed, but we
want to support the case for out-of-tree callers which needs the string
parameter to be persisted.

Change-Id: I371842607f8360e83283b92ff9750f567fa375ca
Signed-off-by: Yash Raj Singh <yash.rajsingh@nutanix.com>
Reviewed-on: https://review.spdk.io/gerrit/c/spdk/spdk/+/25209
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Reviewed-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Konrad Sztyber <konrad.sztyber@intel.com>
Community-CI: Mellanox Build Bot
Community-CI: Broadcom CI <spdk-ci.pdl@broadcom.com>