Port of CI/CD from spdk-community-ci
This removes the gerrit-integration that was done on:

  https://github.com/spdk-community-ci/dispatcher

The CI jobs and workflows need slight adjustments, as they assumed the
SPDK repository was "external." This assumption, along with reporting to
Gerrit, is removed. And in addition:

* Replaced the use of Self-Hosted runners with GitHUB Hosted

  - This is done to evaluate how much functional verification we can
    achieve without adding more capable self-hosted runners
  - Reduced the VM core count from 12 to 8, going down to 4 fails lcore
    tests

* Switched "ci-env" to using ghcr.io/refenv/cijoe-docker:v0.9.47

  - This removes the need to maintain a "spdk-github-ci-docker-image"
  - Adjusted to changes in CIJOE v0.9.47 (eg. guest initialization)
  - By using this, then we leverage the work of CIJOE and other projects
    using it do to nested virtualization e.g. the xNVMe project

* Adjusted the CIJOE workflow to fix overship of github repository

* Refactored the GHA "autorun" workflow

  - Split into parallel runs -- reducing total processing time
  - Re-wrote to avoid redundant descriptions by using "matrix.workflow",
    e.g. parametrized the job-logic

The "autorun.yml" is the only of the provided workflows that have been
"ported", the remaining need a bit more work (qcow image generation,
self-hosted-runner maintenance). But this, can get the ball rolling.. so
to speak.

This runs the tests inside qemu-guests (VMs) with .qcow2 images that
are generated for the task to ensure proper execution environment.
The qemu-guests allows for NVMe device emulation, thereby testing
NVMe Command-sets that are not commonly available such as ZNS, and a
variation of controller parameters such as MDTS size.

Places the CIJOE scripts inside the SPDK "scripts/cijoe" as it seems
like the most appropriate place to put them.

Change-Id: Ic3b9e578a9adfbb3c72166b01ffb8c9b4e306900
Signed-off-by: Simon A. F. Lund <os@safl.dk>