autorun_post: normalize paths in coverage files
Test runs might (and in most cases probably are) executed
on a different entity than the one generating aggregated
coverage report. Modify all paths in the coverage files
so that they match current workspace.

Also remove some similar old replace operation which was
done before running genhtml, and which was searching
for "./repo" in lcov's SF path. We haven't been using
"repo" in checkout path for quite a long time now, so
it's very out dated.

Change-Id: Iba9074d764b0c47e6cbe63a2d8375a78b0ed1074
Signed-off-by: Karol Latecki <karol.latecki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25675
Reviewed-by: Michal Berger <michal.berger@nutanix.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>
Reviewed-by: Tomasz Zawadzki <tomasz@tzawadzki.com>
Tested-by: Tomasz Zawadzki <tomasz@tzawadzki.com>