nvmf: introduced implementation of io cancel
implemented TP4097 for nvmf target
Cancel command was added to the NVMe protocol to improve the Abort mechanism.
For example. Abort commands can be only sent on the Admin queue, and the Admin
queue is limited in size, so we cannot Abort many commands at once. The commands
that needs to be Aborted are in the IO queues, so it might take too long to get
the command on the Admin queue, and look for it in the IO queue. Cancel command
is implemented on the IO queue, and that gives much better chance to actually
being able to catch and Abort. We can also issue as many Cancel commands as the
queue allows for the IO commands

implemented as synchronous flow
added log component io-cancel,
fixed spdk formatting

Change-Id: I7d5448d697fa09f41549b4934aab81c2e13fc55b
Signed-off-by: Leonid Chernin <leonidc@il.ibm.com>
(cherry picked from commit d0b226a9e29e2f7851e691db4d77c52e9d4a4d6e)