nvme: add MSI-X vector count quirk for Micron SSDs
At least some models of Micron SSDs only allocate MSI-X vector count
equal to the number of IO queues - this does not account for the
admin queue. This results in driver initialization failure for the
Micron SSDs when interrupt mode is enabled, since we base the requested
count on the number of IO queues plus one.

So add a quirk to lower the number of supported IO queues by one. Its
not clear how many Micron SSD models out there have this problem, so
we will just apply it to all of them.

Fixes issue #3645.

Signed-off-by: Jim Harris <jim.harris@nvidia.com>
Change-Id: Ibf986c46de156755101aa8f512d290a713666493
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25809
Reviewed-by: Ben Walker <ben@nvidia.com>
Reviewed-by: Aleksey Marchuk <alexeymar@nvidia.com>
Reviewed-by: Changpeng Liu <changpeliu@tencent.com>
Community-CI: Mellanox Build Bot
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: GangCao <gang.cao@intel.com>