bdev/nvme: Cache current for non-optimized path in active-passive
The patch
https://github.com/spdk/spdk/commit/20f78694d4ad49f8cdc8905cefec5dc1383bf6fd
change nvme_io_path_is_current() to return true if io_path is the first
one available on the list in active-passive mode.

This caused a problem when the disable-auto-failback feature was tested.

When the disable-auto-failback feature is enabled, the active available
path should stay current even if the preferred path is back to be available.

However, current was changed because no path was used after cache reset
and the preferred path is located at the head of the list.

The original purpose of the patch was for ANA optimized/non-optimized.

If ANA event is received, I/O path cache is always cleared. Hence, we
can cache ANA non-optimized path in active-passive mode.

Let's do this. Then, a strange behavior for the disable-auto-faiback
feature is fixed and the original purpose is kept.

Change-Id: I12bee2e104b9518ea0c19639164ceca20646248c
Signed-off-by: Shuhei Matsumoto <smatsumoto@nvidia.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/25884
Community-CI: Mellanox Build Bot
Reviewed-by: Ankit Kumar <ankit.kumar@samsung.com>
Reviewed-by: Jim Harris <jim.harris@nvidia.com>
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Ben Walker <ben@nvidia.com>