test/check_so_deps: use latest tag on branch if VERSION tag is missing
Fixes:
(2104eac) test/check_so_deps: use VERSION to look for prior tags

Prior to patch above, the prior version tag was assumed to be latest
on the branch. It has been changed to lookup the prior tag by
VERSION file, which allowed us to create multiple tags (-rc*).

Yet it is now required for the tag matching VERSION to exist on repo.
This is not the case for patches that change VERSION, as they
cannot have a tag before being merged.
To handle this, if tag matching the VERSION does not exist,
fallback to prior behaviour.

Change-Id: Ib19e007a89794306866e827df7a628b26dcadfcb
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@nutanix.com>
Reviewed-on: https://review.spdk.io/c/spdk/spdk/+/26075
Tested-by: SPDK Automated Test System <spdkbot@gmail.com>
Reviewed-by: Mateusz Kozlowski <mateusz.kozlowski@solidigm.com>
Reviewed-by: Konrad Sztyber <ksztyber@nvidia.com>